+++ test/clj/lambdacd/testsupport/data.clj
@@ -4,12 +4,13 @@
 
 
 (defn some-ctx []
-   {:_pipeline-state (atom {})
-    :step-id         [42]
-    :result-channel  (async/chan (async/dropping-buffer 100))
-    :config          {:home-dir (utils/create-temp-dir)}
-    :is-killed       (atom false)
-    :_out-acc        (atom "")})
+  {:_pipeline-state      (atom {})
+   :step-id              [42]
+   :result-channel       (async/chan (async/dropping-buffer 100))
+   :step-results-channel (async/chan (async/dropping-buffer 100))
+   :config               {:home-dir (utils/create-temp-dir)}
+   :is-killed            (atom false)
+   :_out-acc             (atom "")})
 
 (defn some-ctx-with [& args]
   (apply assoc (some-ctx) args))