+++ CHANGELOG.md
@@ -7,14 +7,16 @@ The official release will have a defined and more stable API. If you are already
 
 ## 0.4.0
 
-* Bug fixes
+* Improvements:
+  * Cleaned up status inheritance to make it more consistent
+* Bug fixes:
   * Fix bug in retriggering that left next step in undefined state (#26)
 * API changes:
   * Steps returning no `:status` will now be treated as failures instead of receiving status `:undefined`
   * Removed deprecated `:result-channel` argument for `execute-step`
   * Removed deprecated `core/new-base-context-for`
-  * `core/execute-step` always requires a `:result-channel` in `ctx`.
-    This is already the case when used in conjunction with `execute-steps`
+  * `core/execute-step` does no longer output the result-channel data to a `:result-channel` in ctx.
+    Was replaced with `:step-results-channels` which provides a stream of complete, aggregated step-result data.
 
 ## 0.3.2